*** empty log message ***
git-svn-id: file:///usr/local/svn/fsc2@3528 934aee4a-8510-0410-ae50-ec5b2ea4bb86

+++ trunk/fsc2/src/util.c
@@ -70,7 +70,7 @@ char *get_string( const char *fmt, ... )
 		break;
 	}
 
-	if ( wr + 1 < len )            /* trim string to number of needed chars */
+	if ( ( size_t ) wr + 1 < len )  /* trim string to number of needed chars */
 		T_realloc( c, ( size_t ) wr + 1 );
 
 	return c;
